requirements.txt
telethon==1.28.5
pyrogram==2.0.106
PyQt6==6.4.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2
matplotlib==3.7.1
pandas==2.0.2
loguru==0.7.0
spacy==3.5.3 --no-binary blis,preshedpytesseract==0.3.10
pillow==10.0.0

# Testing
pytest==7.4.0
pytest-asyncio==0.23.5

# Safety
python-dotenv==1.0.0
# Configuration
toml==0.10.2
